Sandbox: make sandbox controller plugin
