+++ api/types/sandbox.proto
@@ -40,12 +40,14 @@ message Sandbox {
 	// Spec is sandbox configuration (kin of OCI runtime spec), spec's data will be written to a config.json file in the
 	// bundle directory (similary to OCI spec).
 	google.protobuf.Any spec = 3;
+	// Sandboxer is the name of the sandbox controller who manages the sandbox.
+	string sandboxer = 4;
 	// Labels provides an area to include arbitrary data on containers.
-	map<string, string> labels  = 4;
+	map<string, string> labels  = 5;
 	// CreatedAt is the time the container was first created.
-	google.protobuf.Timestamp created_at = 5;
+	google.protobuf.Timestamp created_at = 6;
 	// UpdatedAt is the last time the container was mutated.
-	google.protobuf.Timestamp updated_at = 6;
+	google.protobuf.Timestamp updated_at = 7;
 	// Extensions allow clients to provide optional blobs that can be handled by runtime.
-	map<string, google.protobuf.Any> extensions = 7;
+	map<string, google.protobuf.Any> extensions = 8;
 }
